compiler: generate syntax-full.jar, move some files
