+++ syntax/build.gradle.kts
@@ -1,7 +1,9 @@
 // Copyright (c) 2020-2024 Tesla (Yinsen) Zhang.
 // Use of this source code is governed by the MIT license that can be found in the LICENSE.md file.
 
-//CommonTasks.nativeImageConfig(project)
+import org.aya.gradle.CommonTasks
+
+CommonTasks.fatJar(project, "N/A")
 
 dependencies {
   api(project(":tools-kala"))
