+++ syntax/src/main/java/org/aya/syntax/core/pat/BindEater.java
@@ -1,13 +1,11 @@
 // Copyright (c) 2020-2024 Tesla (Yinsen) Zhang.
 // Use of this source code is governed by the MIT license that can be found in the LICENSE.md file.
-package org.aya.tyck.pat;
+package org.aya.syntax.core.pat;
 
 import kala.collection.SeqView;
 import kala.collection.immutable.ImmutableSeq;
 import kala.collection.mutable.MutableList;
 import kala.value.MutableValue;
-import org.aya.syntax.core.pat.Pat;
-import org.aya.syntax.core.pat.PatToTerm;
 import org.aya.syntax.core.term.Term;
 import org.aya.syntax.core.term.call.ConCallLike;
 import org.aya.util.error.Panic;
