+++ syntax/src/main/java/org/aya/syntax/core/pat/PatMatcher.java
@@ -1,21 +1,18 @@
 // Copyright (c) 2020-2024 Tesla (Yinsen) Zhang.
 // Use of this source code is governed by the MIT license that can be found in the LICENSE.md file.
-package org.aya.normalize;
+package org.aya.syntax.core.pat;
 
 import kala.collection.immutable.ImmutableSeq;
 import kala.collection.mutable.FreezableMutableList;
 import kala.collection.mutable.MutableList;
 import kala.control.Result;
 import org.aya.generic.State;
-import org.aya.syntax.core.pat.Pat;
-import org.aya.syntax.core.pat.PatToTerm;
 import org.aya.syntax.core.term.MetaPatTerm;
 import org.aya.syntax.core.term.Term;
 import org.aya.syntax.core.term.TupTerm;
 import org.aya.syntax.core.term.call.ConCall;
 import org.aya.syntax.core.term.call.ConCallLike;
 import org.aya.syntax.core.term.repr.IntegerTerm;
-import org.aya.tyck.pat.BindEater;
 import org.aya.util.error.Panic;
 import org.jetbrains.annotations.NotNull;
